Remove private helper call from proptests
